backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "fastapi>=0.121.3",
    "openai-whisper>=20250625",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "resend>=2.19.0",
    "spoon-ai-sdk>=0.3.3",
    "spoon-toolkits>=0.2.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

